Moves guard function to module
