+++ src/routes/admin/mod.rs
@@ -1,22 +1,12 @@
 pub mod destination;
 pub mod user;
-use crate::db::model::user::User;
 use actix_web::web;
+use crate::guard::admin::admin_guard;
 
 pub fn configure_routes(cfg: &mut web::ServiceConfig) {
   cfg.service(
     web::scope("")
-      .guard(actix_web::guard::fn_guard(|ctx| {
-        let req_data = ctx.req_data();
-        let user_optional = req_data.get::<User>();
-        if user_optional.is_none() {
-          return false;
-        }
-        if user_optional.unwrap().role != "admin" {
-          return false;
-        }
-        return true;
-      }))
+      .guard(actix_web::guard::fn_guard(admin_guard))
       .service(web::scope("/users").configure(user::configure_routes))
       .service(web::scope("/destinations").configure(destination::configure_routes)),
   );
